sync-github-jira/requirements.txt
boto3=1.9.130
jira==2.0.0
PyGithub==1.43.6
